simplepv: change int minUpdatePeriod to long minUpdatePeriodInMs.

+++ core/plugins/org.csstudio.simplepv.pvmanager/src/org/csstudio/simplepv/pvmanager/PVManagerPVFactory.java
@@ -20,7 +20,7 @@ import org.csstudio.simplepv.IPV;
 public class PVManagerPVFactory extends AbstractPVFactory {
 
 	@Override
-	public IPV createPV(String name, boolean readOnly, int minUpdatePeriod, boolean bufferAllValues,
+	public IPV createPV(String name, boolean readOnly, long minUpdatePeriod, boolean bufferAllValues,
 			Executor notificationThread, ExceptionHandler exceptionHandler) {
 		return new PVManagerPV(name, readOnly, minUpdatePeriod, bufferAllValues, notificationThread, exceptionHandler);
 	}
